chore(fluxWizard): refactor fluxPeriodFromRangeSeconds
